+++ parser/src/ast/array_len.rs
@@ -1,5 +1,5 @@
 use super::{LitPosInt, NamedRef};
-use crate::error::{ConstIntNotFound, ExpectedConstIntFoundType};
+use crate::error::{ConstIntNotFound, ExpectedConstIntFoundService, ExpectedConstIntFoundType};
 use crate::grammar::Rule;
 use crate::validate::Validate;
 use crate::Span;
@@ -58,6 +58,7 @@ impl ArrayLenValue {
 
             Self::Ref(ty) => {
                 ConstIntNotFound::validate(ty, validate);
+                ExpectedConstIntFoundService::validate(ty, validate);
                 ExpectedConstIntFoundType::validate(ty, validate);
 
                 ty.validate(validate);