+++ parser/src/error.rs
@@ -9,6 +9,7 @@ mod duplicate_service_uuid;
 mod duplicate_struct_field;
 mod duplicate_struct_field_id;
 mod empty_enum;
+mod expected_const_int_found_service;
 mod expected_const_int_found_type;
 mod expected_type_found_const;
 mod expected_type_found_service;
@@ -42,6 +43,7 @@ pub use duplicate_service_uuid::DuplicateServiceUuid;
 pub use duplicate_struct_field::DuplicateStructField;
 pub use duplicate_struct_field_id::DuplicateStructFieldId;
 pub use empty_enum::EmptyEnum;
+pub use expected_const_int_found_service::ExpectedConstIntFoundService;
 pub use expected_const_int_found_type::ExpectedConstIntFoundType;
 pub use expected_type_found_const::ExpectedTypeFoundConst;
 pub use expected_type_found_service::ExpectedTypeFoundService;
@@ -75,6 +77,7 @@ pub enum Error {
     DuplicateStructField(DuplicateStructField),
     DuplicateStructFieldId(DuplicateStructFieldId),
     EmptyEnum(EmptyEnum),
+    ExpectedConstIntFoundService(ExpectedConstIntFoundService),
     ExpectedConstIntFoundType(ExpectedConstIntFoundType),
     ExpectedTypeFoundConst(ExpectedTypeFoundConst),
     ExpectedTypeFoundService(ExpectedTypeFoundService),
@@ -114,6 +117,7 @@ impl Diagnostic for Error {
             Self::DuplicateStructField(e) => e.schema_name(),
             Self::DuplicateStructFieldId(e) => e.schema_name(),
             Self::EmptyEnum(e) => e.schema_name(),
+            Self::ExpectedConstIntFoundService(e) => e.schema_name(),
             Self::ExpectedConstIntFoundType(e) => e.schema_name(),
             Self::ExpectedTypeFoundConst(e) => e.schema_name(),
             Self::ExpectedTypeFoundService(e) => e.schema_name(),
@@ -149,6 +153,7 @@ impl Diagnostic for Error {
             Self::DuplicateStructField(e) => e.format(parsed),
             Self::DuplicateStructFieldId(e) => e.format(parsed),
             Self::EmptyEnum(e) => e.format(parsed),
+            Self::ExpectedConstIntFoundService(e) => e.format(parsed),
             Self::ExpectedConstIntFoundType(e) => e.format(parsed),
             Self::ExpectedTypeFoundConst(e) => e.format(parsed),
             Self::ExpectedTypeFoundService(e) => e.format(parsed),